back_end/requirements.txt
django==4.2.5
djangorestframework==3.14.0
asgiref==3.7.2
sqlparse==0.4.4